Move `ThemeOption` and `ThemeSelector` to new files
